+++ android/app/build.gradle
@@ -1,102 +1,4 @@
-/*
-def localProperties = new Properties()
-def localPropertiesFile = rootProject.file('local.properties')
-if (localPropertiesFile.exists()) {
-    localPropertiesFile.withReader('UTF-8') { reader ->
-        localProperties.load(reader)
-    }
-}
 
-def flutterRoot = localProperties.getProperty('flutter.sdk')
-if (flutterRoot == null) {
-    throw new FileNotFoundException("Flutter SDK not found. Define location with flutter.sdk in the local.properties file.")
-}
-
-def flutterVersionCode = localProperties.getProperty('flutter.versionCode')
-if (flutterVersionCode == null) {
-    flutterVersionCode = '1'
-}
-
-def flutterVersionName = localProperties.getProperty('flutter.versionName')
-if (flutterVersionName == null) {
-    flutterVersionName = '1.0'
-}
-
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply from: "$flutterRoot/packages/flutter_tools/gradle/flutter.gradle"
-
-android {
-    defaultConfig {
-
-        minSdkVersion 15
-        targetSdkVersion 30
-        multiDexEnabled true
-        renderscriptTargetApi 18
-        renderscriptSupportModeEnabled true
-
-
-    }
-
-    dataBinding {
-        enabled = false
-    }
-
-    dependencies {
-        compile files('libs/renderscript-v8.jar')
-        implementation "androidx.multidex:multidex:2.0.1"
-        def multidex_version = "2.0.1"
-        implementation "androidx.multidex:multidex:$multidex_version"
-        implementation 'com.android.support:multidex:1.0.3'
-    }
-    compileSdkVersion 30
-
-    sourceSets {
-        main.java.srcDirs += 'src/main/kotlin'
-    }
-
-    defaultConfig {
-        // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.example.book_bank"
-        minSdkVersion 16
-        targetSdkVersion 30
-        versionCode flutterVersionCode.toInteger()
-        versionName flutterVersionName
-    }
-
-    buildTypes {
-        release {
-            // TODO: Add your own signing config for the release build.
-            // Signing with the debug keys for now, so `flutter run --release` works.
-            signingConfig signingConfigs.debug
-        }
-    }
-
-}
-
-flutter {
-    source '../..'
-}
-
-repositories {
-    maven {
-        url 'some/path/my_flutter/build/host/outputs/repo'
-        // This is relative to the location of the build.gradle file
-        // if using a relative path.
-    }
-    maven {
-        url 'https://storage.googleapis.com/download.flutter.io'
-    }
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    //debugImplementation 'com.example.flutter_module:flutter_debug:1.0'
-    profileImplementation 'com.example.flutter_module:flutter_profile:1.0'
-    releaseImplementation 'com.example.flutter_module:flutter_release:1.0'
-
-
-}*/
 def localProperties = new Properties()
 def localPropertiesFile = rootProject.file('local.properties')
 if (localPropertiesFile.exists()) {
@@ -170,5 +72,6 @@ dependencies {
     implementation platform('com.google.firebase:firebase-bom:26.7.0')
     implementation "androidx.multidex:multidex:2.0.1"
     implementation 'com.google.firebase:firebase-analytics'
+    implementation 'com.firebaseui:firebase-ui-storage:6.4.0'
     
 }
