+++ lib/main.dart
@@ -8,9 +8,9 @@ import 'login_screen.dart';
 import 'package:flutter_login/flutter_login.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'dart:async';
-//import 'dart:convert';
 import 'Product.dart';
-import 'FirebaseData.dart';
+import 'test.dart';
+import 'load_pdf.dart';
 String userUid = "no user";
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -81,7 +81,7 @@ class Myapp extends DrawerExample {
           bodyText2: TextStyle(fontFamily: 'NatoSans'),
         ),
       ),
-      home: LoginScreen(),
+      home: Test(),
     );
   }
 }
@@ -147,7 +147,7 @@ class DrawerExample extends StatelessWidget {
                   Navigator.push(
                       context,
                       new MaterialPageRoute(
-                          builder: (context) => new DrawerExample()));
+                          builder: (context) => new Test()));
                 },
               ),
               new ListTile(
@@ -170,7 +170,7 @@ class DrawerExample extends StatelessWidget {
                   // Then close the drawer
                   Navigator.pop(context);
                   Navigator.push(
-                      context, new MaterialPageRoute(builder: (context) => new Profile()));
+                      context, new MaterialPageRoute(builder: (context) => new LoadFirbaseStoragePdf()));
                 },
               ),
               new ListTile(
@@ -520,282 +520,3 @@ class contact_us extends StatelessWidget {
   }
 }
 
-/*class MyHomePage extends StatelessWidget {
-  final String title;
-  final Future<List<Product>> products;
-  MyHomePage({Key key, this.title, this.products}) : super(key: key);
-  // final items = Product.getProducts();
-  @override
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(title: Text("Product Navigation")),
-        body: Center(
-          child: StreamBuilder<QuerySnapshot>(
-            stream: products,
-            builder: (context, snapshot) {
-              if (snapshot.hasError) print(snapshot.error);
-              if(snapshot.hasData) {
-                List<DocumentSnapshot> documents = snapshot.data.documents;
-                List<Product> items = List<Product>();
-                for(var i = 0; i < documents.length; i++) {
-                  DocumentSnapshot document = documents[i];
-                  items.add(Product.fromMap(document.data));
-                }
-                return ProductBoxList(items: items);
-              } else {
-                return Center(child: CircularProgressIndicator());
-              }
-            },
-          ),
-        ));
-  }
-
-}
-class ProductPage extends StatelessWidget {
-  ProductPage({Key key, this.item}) : super(key: key);
-  final Product item;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(this.item.name),
-      ),
-      body: Center(
-        child: Container(
-          padding: EdgeInsets.all(0),
-          child: Column(
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: <Widget>[
-                Image.asset("assets/" + this.item.image),
-                Expanded(
-                    child: Container(
-                        padding: EdgeInsets.all(5),
-                        child: Column(
-                          mainAxisAlignment:
-                          MainAxisAlignment.spaceEvenly,
-                          children: <Widget>[
-                            Text(this.item.name,
-                                style: TextStyle(fontWeight:
-                                FontWeight.bold)),
-                            Text(this.item.description),
-                            Text("Price: " +
-                                this.item.price.toString()),
-                            RatingBox(),
-                          ],
-                        )))
-              ]),
-        ),
-      ),
-    );
-  }
-}
-class RatingBox extends StatefulWidget {
-  @override
-  _RatingBoxState createState() => _RatingBoxState();
-}
-class _RatingBoxState extends State<RatingBox> {
-  int _rating = 0;
-  void _setRatingAsOne() {
-    setState(() {
-      _rating = 1;
-    });
-  }
-  void _setRatingAsTwo() {
-    setState(() {
-      _rating = 2;
-    });
-  }
-  void _setRatingAsThree() {
-    setState(() {
-      _rating = 3;
-    });
-  }
-  Widget build(BuildContext context) {
-    double _size = 20;
-    print(_rating);
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.end,
-      crossAxisAlignment: CrossAxisAlignment.end,
-      mainAxisSize: MainAxisSize.max,
-      children: <Widget>[
-        Container(
-          padding: EdgeInsets.all(0),
-          child: IconButton(
-            icon: (_rating >= 1
-                ? Icon(
-              Icons.star,
-              size: _size,
-            )
-                : Icon(
-              Icons.star_border,
-              size: _size,
-            )),
-            color: Colors.red[500],
-            onPressed: _setRatingAsOne,
-            iconSize: _size,
-          ),
-        ),
-        Container(
-          padding: EdgeInsets.all(0),
-          child: IconButton(
-            icon: (_rating >= 2
-                ? Icon(
-              Icons.star,
-              size: _size,
-            )
-                : Icon(
-              Icons.star_border,
-              size: _size,
-            )),
-            color: Colors.red[500],
-            onPressed: _setRatingAsTwo,
-            iconSize: _size,
-          ),
-        ),
-        Container(
-          padding: EdgeInsets.all(0),
-          child: IconButton(
-            icon: (_rating >= 3
-                ? Icon(
-              Icons.star,
-              size: _size,
-            )
-                : Icon(
-              Icons.star_border,
-              size: _size,
-            )),
-            color: Colors.red[500],
-            onPressed: _setRatingAsThree,
-            iconSize: _size,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
- */
-/*class FirstPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: new AppBar(
-        leading: IconButton(
-            icon: Icon(Icons.arrow_back, color: Colors.black),
-            onPressed: () => Navigator.push(context,
-                MaterialPageRoute(builder: (context) => DrawerExample()))),
-        title: Text("all books"),
-        centerTitle: true,
-      ),
-      body: new ListView(
-        shrinkWrap: true,
-        padding: const EdgeInsets.fromLTRB(2.0, 10.0, 2.0, 10.0),
-        children: <Widget>[
-          ProductBox(
-              name: "C++",
-              description: "book for learning basic C++ programing language",
-              price: 1000,
-              image: "c++.png"),
-          ProductBox(
-              name: "java",
-              description: "book for learning basic java programing language",
-              price: 800,
-              image: "java.png"),
-          ProductBox(
-              name: "java 8",
-              description:
-                  "book for learning basic to advance java 8 programing language",
-              price: 2000,
-              image: "java 8.png"),
-          ProductBox(
-              name: "let us c",
-              description: "book for learning basic c programing language",
-              price: 1500,
-              image: "LetUsC.png"),
-          ProductBox(
-              name: "python",
-              description: "book for learning basic python programing language",
-              price: 100,
-              image: "python.png"),
-          ProductBox(
-              name: "coding",
-              description: book for learning basic C++ programing language,
-              price: 20,
-              image: "coders at work.png"),
-        ],
-      ),
-    );
-  }
-}
-*/
-
-
-/*
-class ProductBox extends StatelessWidget {
-  ProductBox({Key key, this.item}) : super(key: key);
-  final Product item;
-  Widget build(BuildContext context) {
-    return Container(
-        padding: EdgeInsets.all(2),
-        height: 140,
-        child: Card(
-          child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: <Widget>[
-                Image.asset("assets/" + this.item.image),
-                Expanded(
-                    child: Container(
-                        padding: EdgeInsets.all(5),
-                        child: Column(
-                          mainAxisAlignment:
-                          MainAxisAlignment.spaceEvenly,
-                          children: <Widget>[
-                            Text(this.item.name,
-                                style: TextStyle(fontWeight:
-                                FontWeight.bold)),
-                            Text(this.item.description),
-                            Text("Price: " + this.item.price.toString()),
-                            RatingBox(),
-                          ],
-                        )))
-              ]),
-        ));
-  }
-}
-*/
-/*class ProductBox extends StatelessWidget {
-  ProductBox({Key key, this.name, this.description, this.price, this.image})
-      : super(key: key);
-  final String name;
-  final String description;
-  final int price;
-  final String image;
-  Widget build(BuildContext context) {
-    return Container(
-        padding: EdgeInsets.all(2),
-        height: 120,
-        child: Card(
-            child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: <Widget>[
-              Image.asset("assets/" + image),
-              Expanded(
-                  child: Container(
-                      padding: EdgeInsets.all(5),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        children: <Widget>[
-                          Text(this.name,
-                              style: TextStyle(
-                                  fontWeight: FontWeight.bold,
-                                  color: Colors.red)),
-                          Text(this.description),
-                          Text("Price: ₹ " + this.price.toString()),
-                        ],
-                      )))
-            ])));
-  }
-}
-*/
