--- lib/product.dart
@@ -1,15 +1,0 @@
-class Product {
-  final String name;
-  final String description;
-  final int price;
-  final String image;
-  Product(this.name, this.description, this.price, this.image);
-  factory Product.fromMap(Map<String, dynamic> json) {
-    return Product(
-      json['name'],
-      json['description'],
-      json['price'],
-      json['image'],
-    );
-  }
-}
