+++ pubspec.yaml
@@ -31,11 +31,14 @@ dependencies:
   cloud_firestore: ^1.0.1
   flutter_search_bar: ^2.1.0
   english_words: ^4.0.0
-  provider: ^4.0.2
+  provider: ^5.0.0
   flutter_login: ^1.0.14
   shared_preferences: ^2.0.4
   modal_progress_hud: ^0.1.3
   http: ^0.13.0
+  flutter_full_pdf_viewer: ^1.0.6
+  file_picker: ^3.0.1
+  firebase_storage: ^8.0.1
   #dio: ^3.0.9
   #path_provider: ^1.6.5
   #downloads_path_provider_28: ^0.1.0
@@ -57,7 +60,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_launcher_icons: ^0.9.0
-  gradient_widgets: ^0.5.1
+  gradient_widgets: ^0.6.0
 flutter_icons:
     ios: true
     android: true
